% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{summary.aei_tbl}
\alias{summary.aei_tbl}
\title{Summary method for aei_tbl}
\usage{
\method{summary}{aei_tbl}(object, ...)
}
\arguments{
\item{object}{An \code{aei_tbl}.}

\item{...}{Passed to the underlying \code{summary.data.frame} method.}
}
\value{
Invisibly returns the standard data frame summary.
}
\description{
Summary method for aei_tbl
}
